refactor(marketing): encapsulate podcast page
CodeTether-Provenance-ID: ctprov_03b53e9010bc47638b9296a355f470c0
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: 5ed11d33-e305-492e-a365-361adb6d4a03
